Fixed 500 response while POST /play../puzzle.
It was about the TrainingRankings is not created on start [user registration].

When we were trying to get TrainingRanking, we create new as fallback.

And this cause an error, when too much request is sent. In database the TrainingRanking is created, but second request also trying to insert with same data.